+++ src/copy_tree.rs
@@ -14,6 +14,8 @@ pub struct CopyOptions {
 }
 
 /// Copy files and other entries from one tree to another.
+///
+/// NOTE: Although this is public, it's suggested to use `Archive::backup` if possible.
 pub fn copy_tree<ST: ReadTree, DT: WriteTree>(
     source: &ST,
     mut dest: DT,