+++ NEWS.md
@@ -37,6 +37,8 @@ consistent:
   compatible versions. (At present, everything from the 0.6 series that 
   changed the format.)
 
+- New simpler `Archive::backup` public API.
+
 ## v0.6.3 2020-05-30
 
 ### Performance improvements